refactor(arch): Merge statistics_utils and update_statistic into utils.statistics
