+++ app/background_processes/__init__.py
@@ -1,2 +1,1 @@
-from app.background_processes import update_statistic  # noqa: F401 #isort:skip
 from app.background_processes import fetcher  # noqa: F401 #isort:skip
