+++ app/utils/statistics.py
@@ -1,15 +1,144 @@
 import logging
+from contextlib import contextmanager
+from datetime import datetime
 
-from sqlalchemy import Integer, and_, distinct, func, or_, text
+from sqlalchemy import Date, Integer, and_, cast, distinct, func, or_, text
+from sqlalchemy.orm import Session
 
-from app.models import Application, ApplicationStatus  # noqa: F401 # isort:skip
-from app.models import Borrower, CreditProduct  # noqa: F401 # isort:skip
-from app.models import CreditType, StatisticData, Lender  # noqa: F401 # isort:skip
+from app.db import get_db, transaction_session_logger
+from app.models import (
+    Application,
+    ApplicationStatus,
+    Borrower,
+    CreditProduct,
+    CreditType,
+    Lender,
+    Statistic,
+    StatisticData,
+    StatisticType,
+)
 
 logger = logging.getLogger(__name__)
 
 
-def get_base_query(sessionBase, start_date, end_date, lender_id):
+def update_statistics(db_provider: Session = get_db):
+    """
+    Update and store various statistics related to applications and lenders in the database.
+
+    This function retrieves and logs different types of statistics related to applications
+    and lenders. It uses the `get_general_statistics`, `get_msme_opt_in_stats`,
+    and `get_count_of_fis_choosen_by_msme` functions
+    to fetch the respective statistics. The retrieved statistics are then logged using
+    the `logger.info()` function.
+
+    After fetching the general statistics, this function attempts to store them in the database
+    as an instance of the `Statistic` model. The statistics are stored with the type set to
+    `StatisticType.APPLICATION_KPIS`. The `Statistic` model contains a JSON field to store
+    the actual statistical data.
+
+    If an error occurs during the process, it is caught and logged using `logger.exception()`.
+    The database session is rolled back in case of an exception to prevent any changes from
+    being committed to the database.
+
+    Note:
+    - The function utilizes the `get_db()` context manager to open a database session.
+
+    Example usage:
+    >>> update_statistics()
+    """
+
+    with contextmanager(db_provider)() as session:
+        with transaction_session_logger(session, "Error saving statistics"):
+            # Get general Kpis
+            statistic_kpis = get_general_statistics(session, None, None, None)
+            # Try to get the existing row
+            statistic_kpi_data = (
+                session.query(Statistic)
+                .filter(
+                    cast(Statistic.created_at, Date) == datetime.today().date(),
+                    Statistic.type == StatisticType.APPLICATION_KPIS,
+                )
+                .first()
+            )
+
+            # If it exists, update it
+            if statistic_kpi_data:
+                statistic_kpi_data.data = statistic_kpis
+            # If it doesn't exist, create a new one
+            else:
+                statistic_kpi_data = Statistic(
+                    type=StatisticType.APPLICATION_KPIS,
+                    data=statistic_kpis,
+                    created_at=datetime.now(),
+                )
+                session.add(statistic_kpi_data)
+
+            # Get Opt in statistics
+            statistics_msme_opt_in = get_msme_opt_in_stats(session)
+            statistics_msme_opt_in["sector_statistics"] = [
+                data.dict() for data in statistics_msme_opt_in["sector_statistics"]
+            ]
+            statistics_msme_opt_in["rejected_reasons_count_by_reason"] = [
+                data.dict() for data in statistics_msme_opt_in["rejected_reasons_count_by_reason"]
+            ]
+            statistics_msme_opt_in["fis_choosen_by_msme"] = [
+                data.dict() for data in statistics_msme_opt_in["fis_choosen_by_msme"]
+            ]
+            # Try to get the existing row
+            statistic_opt_data = (
+                session.query(Statistic)
+                .filter(
+                    cast(Statistic.created_at, Date) == datetime.today().date(),
+                    Statistic.type == StatisticType.MSME_OPT_IN_STATISTICS,
+                )
+                .first()
+            )
+
+            # If it exists, update it
+            if statistic_opt_data:
+                statistic_opt_data.data = statistics_msme_opt_in
+            # If it doesn't exist, create a new one
+            else:
+                statistic_opt_data = Statistic(
+                    type=StatisticType.MSME_OPT_IN_STATISTICS,
+                    data=statistics_msme_opt_in,
+                    created_at=datetime.now(),
+                )
+                session.add(statistic_opt_data)
+
+            # Get general Kpis for every lender
+            lender_ids = [id[0] for id in session.query(Lender.id).all()]
+            for lender_id in lender_ids:
+                # Get statistics for each lender
+                statistic_kpis = get_general_statistics(session, None, None, lender_id)
+
+                # Try to get the existing row
+                statistic_kpi_data = (
+                    session.query(Statistic)
+                    .filter(
+                        cast(Statistic.created_at, Date) == datetime.today().date(),
+                        Statistic.type == StatisticType.APPLICATION_KPIS,
+                        Statistic.lender_id == lender_id,
+                    )
+                    .first()
+                )
+
+                # If it exists, update it
+                if statistic_kpi_data:
+                    statistic_kpi_data.data = statistic_kpis
+                # If it doesn't exist, create a new one
+                else:
+                    statistic_kpi_data = Statistic(
+                        type=StatisticType.APPLICATION_KPIS,
+                        data=statistic_kpis,
+                        lender_id=lender_id,
+                        created_at=datetime.now(),
+                    )
+
+                session.add(statistic_kpi_data)
+
+
+def _get_base_query(sessionBase, start_date, end_date, lender_id):
     """
     Create the base query for filtering applications based on the provided start_date, end_date, and lender_id.
 
@@ -74,14 +203,14 @@ def get_general_statistics(session, start_date=None, end_date=None, lender_id=No
     """
 
     # received--------
-    applications_received_query = get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
+    applications_received_query = _get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
         Application.borrower_submitted_at.isnot(None),
     )
 
     applications_received_count = applications_received_query.count()
 
     # approved-------
-    applications_approved_query = get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
+    applications_approved_query = _get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
         or_(
             Application.status == ApplicationStatus.APPROVED,
             Application.status == ApplicationStatus.CONTRACT_UPLOADED,
@@ -92,21 +221,21 @@ def get_general_statistics(session, start_date=None, end_date=None, lender_id=No
     applications_approved_count = applications_approved_query.count()
 
     # rejected--------
-    applications_rejected_query = get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
+    applications_rejected_query = _get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
         Application.status == ApplicationStatus.REJECTED,
     )
 
     applications_rejected_count = applications_rejected_query.count()
 
     # waiting---------
-    applications_waiting_query = get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
+    applications_waiting_query = _get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
         Application.status == ApplicationStatus.INFORMATION_REQUESTED,
     )
 
     applications_waiting_count = applications_waiting_query.count()
 
     # in progress---------
-    applications_in_progress_query = get_base_query(
+    applications_in_progress_query = _get_base_query(
         session.query(Application), start_date, end_date, lender_id
     ).filter(
         or_(
@@ -118,7 +247,7 @@ def get_general_statistics(session, start_date=None, end_date=None, lender_id=No
     applications_in_progress_count = applications_in_progress_query.count()
 
     # credit disbursed---------
-    applications_with_credit_disbursed = get_base_query(
+    applications_with_credit_disbursed = _get_base_query(
         session.query(Application), start_date, end_date, lender_id
     ).filter(
         Application.status == ApplicationStatus.COMPLETED,
@@ -133,7 +262,7 @@ def get_general_statistics(session, start_date=None, end_date=None, lender_id=No
         proportion_of_disbursed = int(applications_with_credit_disbursed_count / applications_approved_count * 100)
 
     # Average amount requested
-    average_amount_requested_query = get_base_query(
+    average_amount_requested_query = _get_base_query(
         session.query(func.avg(Application.amount_requested)),
         start_date,
         end_date,
@@ -149,7 +278,7 @@ def get_general_statistics(session, start_date=None, end_date=None, lender_id=No
 
     # Average Repayment Period
     average_repayment_period_query = (
-        get_base_query(
+        _get_base_query(
             session.query(func.avg(Application.repayment_years * 12 + Application.repayment_months).cast(Integer)),
             start_date,
             end_date,
@@ -167,14 +296,14 @@ def get_general_statistics(session, start_date=None, end_date=None, lender_id=No
     average_repayment_period = average_repayment_period_query.scalar() or 0
 
     # Overdue Application
-    applications_overdue_query = get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
+    applications_overdue_query = _get_base_query(session.query(Application), start_date, end_date, lender_id).filter(
         Application.overdued_at.isnot(None),
     )
 
     applications_overdue_count = applications_overdue_query.count()
 
     # average time to process application
-    average_processing_time_query = get_base_query(
+    average_processing_time_query = _get_base_query(
         session.query(func.avg(Application.completed_in_days)),
         start_date,
         end_date,
@@ -187,7 +316,7 @@ def get_general_statistics(session, start_date=None, end_date=None, lender_id=No
     average_processing_time = int(average_processing_time_result) if average_processing_time_result is not None else 0
     #  get_proportion_of_submited_out_of_opt_in
     application_accepted_query = (
-        get_base_query(session.query(Application), start_date, end_date, lender_id)
+        _get_base_query(session.query(Application), start_date, end_date, lender_id)
         .filter(Application.borrower_submitted_at.isnot(None))
         .count()
     )
